Python Script To Download Images on Webpage
This script was made because I wouldn't install the ImageScraper Python Module. What this script allows the user to do is to download all detected images on appropriate URL. Simply enter the URL and Folder name and the script handles the rest. Currently Accounts for .jpg and .png will add more file formats later on.